+++ RELEASE_NOTES.md
@@ -60,8 +60,10 @@ If you have persisted `KeyboardTheme` values in your app, please reach out via `
 
 ### 🗑️ Deprecations
 
+* `ExternalAutocompleteProvider` is renamed to `RemoteAutocompleteProvider`.
 * `KeyboardContext.tempIsPreviewMode` is no longer used.
 * `KeyboardLocale` sort with insert first parameter is replaced by atomic operations.
+* `StandardAutocompleteProvider` is renamed to `LocalAutocompleteProvider`.
 
 ### 🐛 Bug fixes